fix: refactor remote config tasks
