+++ src/commands/network.ts
@@ -26,7 +26,6 @@ import * as helpers from '../core/helpers.js';
 import {addDebugOptions, resolveValidJsonFilePath, validatePath} from '../core/helpers.js';
 import path from 'path';
 import fs from 'fs';
-import {RemoteConfigTasks} from '../core/config/remote/remote_config_tasks.js';
 import {type KeyManager} from '../core/key_manager.js';
 import {type PlatformInstaller} from '../core/platform_installer.js';
 import {type ProfileManager} from '../core/profile_manager.js';
