+++ src/core/config/remote/remote_config_manager.ts
@@ -215,55 +215,26 @@ export class RemoteConfigManager {
     await self.save();
   }
 
-  /**
-   * Builds a listr task for loading the remote configuration.
-   *
-   * @param argv - arguments containing command input for historical reference.
-   * @returns a Listr task which loads the remote configuration.
-   */
-  public buildLoadTask(argv: {_: string[]}): SoloListrTask<EmptyContextConfig> {
+  public async createAndValidate(cluster: Cluster, context: Context, namespace: Namespace) {
     const self = this;
+    self.k8.setCurrentContext(context);
 
-    return {
-      title: 'Load remote config',
-      task: async (_, task): Promise<void> => {
-        await self.loadAndValidate(argv);
-      },
-    };
-  }
+    if (!(await self.k8.hasNamespace(namespace))) {
+      await self.k8.createNamespace(namespace);
+    }
 
-  /**
-   * Builds a task for creating a new remote configuration, intended for use with Listr task management.
-   * Merges cluster mappings from the provided context into the local configuration, then creates the remote config.
-   *
-   * @returns a Listr task which creates the remote configuration.
-   */
-  public buildCreateTask(cluster: Cluster, context: Context, namespace: Namespace): SoloListrTask<ListrContext> {
-    const self = this;
+    const localConfigExists = this.localConfig.configFileExists();
+    if (!localConfigExists) {
+      throw new SoloError("Local config doesn't exist");
+    }
 
-    return {
-      title: `Create remote config in cluster: ${cluster}`,
-      task: async (_, task): Promise<void> => {
-        self.k8.setCurrentContext(context);
+    self.unload();
+    if (await self.load()) {
+      self.logger.showUser(chalk.red('Remote config already exists'));
+      throw new SoloError('Remote config already exists');
+    }
 
-        if (!(await self.k8.hasNamespace(namespace))) {
-          await self.k8.createNamespace(namespace);
-        }
-
-        const localConfigExists = this.localConfig.configFileExists();
-        if (!localConfigExists) {
-          throw new SoloError("Local config doesn't exist");
-        }
-
-        self.unload();
-        if (await self.load()) {
-          task.title = `${task.title} - ${chalk.red('Remote config already exists')}}`;
-          throw new SoloError('Remote config already exists');
-        }
-
-        await self.create();
-      },
-    };
+    await self.create();
   }
 
   /* ---------- Utilities ---------- */
